// src/orchestrator/type.ts
type ContentPart = TextContent | ImageContentPart;

export type ToolMessage = {
  role: "tool";
  content: string;
  tool_call_id: string;
  name?: string;
};
export type UserMessage = {
  role: "user";
  // ContentParts are only for the "user" role:
  content: string | ContentPart[];
  name?: string;
};

export type AssistantMessage = {
  role: "assistant";
  // ContentParts are only for the "user" role:
  content?: string;
  name?: string;
  tool_calls?: ToolCall[];
};
export type Message =
  | UserMessage
  | AssistantMessage
  | {
      role: "system";
      content: string;
      // If "name" is included, it will be prepended like this
      // for non-OpenAI models: `{name}: {content}`
      name?: string;
    }
  | ToolMessage;
// ...
type FunctionCall = {
  name: string;
  arguments: string;
};

export type ToolCall = {
  id: string;
  type: "function";
  function: FunctionCall;
};